analysis/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.3
click==8.1.6
dnspython==2.4.2
fastapi==0.101.1
h11==0.14.0
idna==3.4
matplotlib==3.7.1
mongoengine==0.27.0
motor==3.2.0
numpy==1.24.2
pandas==1.5.3
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzlocal==5.0.1
uvicorn==0.23.2
